fix sources qui marchaient plus

+++ source.h
@@ -10,16 +10,16 @@ class source
 {
 public:
     source();
-    virtual float ve(float t);
+    virtual double ve(double t);
 };
 
 class source_periodique: public source {
     public:
         source_periodique();
     protected:
-        float frequence;
-        float amplitude;
-        float phase;
+        double frequence;
+        double amplitude;
+        double phase;
 };
 
 class source_aperiodique: public source {
@@ -30,48 +30,48 @@ public:
 class source_sin: public source_periodique {
 public:
     source_sin();
-    float ve(float t);
+    double ve(double t);
 
 };
 
 class source_triangulaire : public source_periodique{
 public:
     source_triangulaire();
-    float ve(float t);
+    double ve(double t);
 private:
-    float amplitude;
-    float phase;
-    float frequence;
+    double amplitude;
+    double phase;
+    double frequence;
 };
 
 class creneau : public source_periodique{
 public:
     creneau();
-    float ve(float t);
+    double ve(double t);
 private:
-    float amplitude;
-    float phase;
-    float frequence;
-    float rapport_cyclique;
+    double amplitude;
+    double phase;
+    double frequence;
+    double rapport_cyclique;
 };
 
 class source_rectangulaire: public source_aperiodique{
 public:
     source_rectangulaire();
-    float ve(float t);
+    double ve(double t);
 private:
-    float amplitude;
-    float phase;
-    float duree;
+    double amplitude;
+    double phase;
+    double duree;
 };
 
 class echelon: public source_aperiodique{
 public:
     echelon();
-    float ve(float t);
+    double ve(double t);
 private:
-    float amplitude;
-    float phase;
+    double amplitude;
+    double phase;
 };
 
 #endif // SOURCE_H
